@@ -1,32 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Immutable;
 using DotVVM.Framework.Compilation.Parser.Binding.Parser;
+using DotVVM.Framework.Compilation.Parser.Dothtml.Parser;
 
 namespace DotVVM.Framework.Compilation.ControlTree.Resolved
 {
     public abstract class ResolvedTypeSpecificationDirective : ResolvedDirective, IAbstractTypeSpecificationDirective
     {
         public BindingParserNode NameSyntax { get; }
-        public ResolvedTypeDescriptor ResolvedType { get; }
+        public ResolvedTypeDescriptor? ResolvedType { get; }
 
-        ITypeDescriptor IAbstractTypeSpecificationDirective.ResolvedType => ResolvedType;
+        ITypeDescriptor? IAbstractTypeSpecificationDirective.ResolvedType => ResolvedType;
 
-        public ResolvedTypeSpecificationDirective(BindingParserNode nameSyntax, ResolvedTypeDescriptor resolvedType)
+        public ResolvedTypeSpecificationDirective(
+            DirectiveCompilationService compilationService,
+            DothtmlDirectiveNode dothtmlNode,
+            BindingParserNode nameSyntax,
+            ImmutableList<NamespaceImport> imports)
+            : base(dothtmlNode)
         {
-            this.NameSyntax = nameSyntax;
-            this.ResolvedType = resolvedType;
+            NameSyntax = nameSyntax;
+            ResolvedType = compilationService.ResolveType(dothtmlNode, nameSyntax, imports);
         }
     }
     public sealed class ResolvedViewModelDirective : ResolvedTypeSpecificationDirective, IAbstractViewModelDirective
     {
-        public ResolvedViewModelDirective(BindingParserNode nameSyntax, ResolvedTypeDescriptor resolvedType)
-            : base(nameSyntax, resolvedType) { }
+        public ResolvedViewModelDirective(
+            DirectiveCompilationService directiveCompilationService,
+            DothtmlDirectiveNode dothtmlNode,
+            BindingParserNode nameSyntax,
+            ImmutableList<NamespaceImport> imports)
+            : base(directiveCompilationService, dothtmlNode, nameSyntax, imports) {
+
+        }
     }
     public sealed class ResolvedBaseTypeDirective : ResolvedTypeSpecificationDirective, IAbstractBaseTypeDirective
     {
-        public ResolvedBaseTypeDirective(BindingParserNode nameSyntax, ResolvedTypeDescriptor resolvedType)
-            : base(nameSyntax, resolvedType) { }
+        public ResolvedBaseTypeDirective(
+            DirectiveCompilationService directiveCompilationService,
+            DothtmlDirectiveNode dothtmlNode,
+            BindingParserNode nameSyntax,
+            ImmutableList<NamespaceImport> imports)
+            : base(directiveCompilationService, dothtmlNode, nameSyntax, imports) {
+
+        }
     }
 }
