@@ -92,7 +92,13 @@ namespace DotVVM.Framework.Controls
             else
                 return null;
         }
+
         /// Sets an expected data context type
-        public static void SetDataContextType(this DotvvmBindableObject obj, DataContextStack? stack) => obj.properties.Set(Internal.DataContextTypeProperty, stack);
+        public static TControl SetDataContextType<TControl>(this TControl control, DataContextStack? stack)
+            where TControl : DotvvmBindableObject
+        {
+            control.properties.Set(Internal.DataContextTypeProperty, stack);
+            return control;
+        }
     }
 }
