@@ -1,11 +1,10 @@
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using DotVVM.Framework.Binding;
 using DotVVM.Framework.Compilation;
 using DotVVM.Framework.Compilation.Binding;
 using DotVVM.Framework.Compilation.ControlTree;
 using DotVVM.Framework.Compilation.ControlTree.Resolved;
+using DotVVM.Framework.Compilation.Directives;
 using DotVVM.Framework.Compilation.Javascript;
 using DotVVM.Framework.Compilation.Styles;
 using DotVVM.Framework.Compilation.Validation;
@@ -66,12 +65,14 @@ namespace Microsoft.Extensions.DependencyInjection
             services.TryAddSingleton<IControlBuilderFactory, DefaultControlBuilderFactory>();
             services.TryAddSingleton<IControlResolver, DefaultControlResolver>();
             services.TryAddSingleton<IControlTreeResolver, DefaultControlTreeResolver>();
+            services.TryAddSingleton<IMarkupDirectiveCompilerPipeline, MarkupDirectiveCompilerPipeline>();
             services.TryAddSingleton<IAbstractTreeBuilder, ResolvedTreeBuilder>();
             services.TryAddSingleton<Func<ControlUsageValidationVisitor>>(s => () => ActivatorUtilities.CreateInstance<ControlUsageValidationVisitor>(s));
             services.TryAddSingleton<IViewCompiler, DefaultViewCompiler>();
             services.TryAddSingleton<IBindingCompiler, BindingCompiler>();
             services.TryAddSingleton<IBindingExpressionBuilder, BindingExpressionBuilder>();
             services.TryAddSingleton<BindingCompilationService, BindingCompilationService>();
+            services.TryAddSingleton<DirectiveCompilationService, DirectiveCompilationService>();
             services.TryAddSingleton<DataPager.CommonBindings>();
             services.TryAddSingleton<IControlUsageValidator, DefaultControlUsageValidator>();
             services.TryAddSingleton<ILocalResourceUrlManager, LocalResourceUrlManager>();
@@ -81,7 +82,14 @@ namespace Microsoft.Extensions.DependencyInjection
             services.TryAddSingleton<IHttpRedirectService, DefaultHttpRedirectService>();
             services.TryAddSingleton<IExpressionToDelegateCompiler, DefaultExpressionToDelegateCompiler>();
 
+            services.AddScoped<IRequestTracer>(s => {
+                var config = s.GetRequiredService<DotvvmConfiguration>();
+                return (config.Diagnostics.PerfWarnings.IsEnabled ? (IRequestTracer)s.GetService<PerformanceWarningTracer>() : null) ?? NullRequestTracer.Instance;
+            });
+            services.TryAddSingleton<JsonSizeAnalyzer>();
+            services.TryAddScoped<PerformanceWarningTracer>();
             services.TryAddScoped<RuntimeWarningCollector>();
+            services.AddTransient<IDotvvmWarningSink, AspNetCoreLoggerWarningSink>();
             services.TryAddScoped<AggregateRequestTracer, AggregateRequestTracer>();
             services.TryAddScoped<ResourceManager, ResourceManager>();
             services.TryAddSingleton<StaticCommandMethodTranslator>();
@@ -117,6 +125,7 @@ namespace Microsoft.Extensions.DependencyInjection
             services.TryAddSingleton<DotvvmErrorPageRenderer>();
             services.AddSingleton<IDotvvmViewCompilationService, DotvvmViewCompilationService>();
             services.AddSingleton<CompilationPageApiPresenter>();
+
 
             return services;
         }
