@@ -3,6 +3,7 @@ import { deserialize } from '../serialization/deserialize'
 import { serialize } from '../serialization/serialize'
 import { serializeDate } from '../serialization/date'
 import { tryCoerce } from '../metadata/coercer';
+import { createComplexObservableSubViewmodel, createComplexObservableViewmodel, ObservableHierarchy, ObservableSubHierarchy } from "./observableHierarchies"
 
 jest.mock("../metadata/typeMap", () => ({
     getTypeInfo(typeId: string) {
@@ -1249,32 +1250,6 @@ function createComplexObservableTargetWithMissingSubHierarchy(): KnockoutObserva
     })
 }
 
-function createComplexObservableViewmodel(): ObservableHierarchy {
-    return {
-        $type: ko.observable("t5"),
-        Prop1: ko.observable("aa"),
-        Prop2: ko.observable(createComplexObservableSubViewmodel())
-    }
-}
-
-function createComplexObservableSubViewmodel(): ObservableSubHierarchy {
-    return {
-        $type: ko.observable("t5_a"),
-        Prop21: ko.observable("bb"),
-        Prop22: ko.observable("cc"),
-        Prop23: ko.observableArray([
-            ko.observable({
-                $type: ko.observable("t5_a_a"),
-                Prop231: ko.observable("dd")
-            }),
-            ko.observable({
-                $type: ko.observable("t5_a_a"),
-                Prop231: ko.observable("ee")
-            })
-        ])
-    }
-}
-
 function createComplexNonObservableViewmodel() {
     return {
         $type: "t5",
@@ -1365,22 +1340,6 @@ class TestData {
         const prop2 = assertObservable(object.Prop2)
         expect(prop2).toBe("bb")
     }
-}
-
-interface ObservableHierarchy {
-    $type: string | KnockoutObservable<string>
-    Prop1: KnockoutObservable<string>
-    Prop2: null | KnockoutObservable<ObservableSubHierarchy>
-}
-
-interface ObservableSubHierarchy {
-    $type: string | KnockoutObservable<string>
-    Prop21: KnockoutObservable<string>
-    Prop22: KnockoutObservable<string>
-    Prop23: KnockoutObservableArray<null | KnockoutObservable<{ 
-        Prop231: null | KnockoutObservable<null | string>,
-        $type: string | KnockoutObservable<string>
-    }>>
 }
 
 const testTypeMap: TypeMap = {
