@@ -2,10 +2,13 @@ import typescript from '@rollup/plugin-typescript'
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import replace from '@rollup/plugin-replace';
+import sveltePreprocess from 'svelte-preprocess';
+import svelte from 'rollup-plugin-svelte';
+import ts from 'typescript'
 //import livereload from '@rollup/plugin-livereload';
 //import { terser } from '@rollup/plugin-terser';
 const production = !process.env.ROLLUP_WATCH;
-export default {
+export default [{
     input: './Scripts/react/react-app.tsx',
     output: {
         format: 'esm',
@@ -14,7 +17,8 @@ export default {
     },
     plugins: [
         typescript({
-            tsconfig: "tsconfig.react.json"
+            tsconfig: "tsconfig.react.json",
+            typescript: ts
         }),
         resolve({ browser: true }),
         commonjs({
@@ -23,10 +27,33 @@ export default {
         replace({
             'process.env.NODE_ENV': JSON.stringify('production')
         })
-        // Watch the `public` directory and refresh the
-        // browser on changes when not in production
-        //!production && livereload('public'),
-        //production && terser(),
     ]
-}
+},{
+    input: './Scripts/svelte/svelte-app.ts',
+    output: {
+        format: 'esm',
+        file: './script/svelte-app.js',
+        sourcemap: !production
+    },
+    plugins: [
+        svelte({
+            dev: !production,
+            css: css => {
+                css.write("svelte-app.css");
+            },
+            preprocess: sveltePreprocess(),
+        }),
+        resolve({ browser: true, dedupe: ['svelte'] }),
+        commonjs({
+            include: 'node_modules/**'
+        }),
+        typescript({
+            tsconfig: "tsconfig.svelte.json",
+            typescript: ts
+        }),
+        replace({
+            'process.env.NODE_ENV': JSON.stringify('production')
+        })
+    ]
+}]
 
